fix(router): Task 17b review round 1 — S1-S8, N1-N4, and controller ruling AE
S1 — three *live* documents still called the `ripupPassNo >= 2` divergence open,
with the diagnosis this task disproved. `crates/fr-router/README.md`'s section is
rewritten as "CLOSED (ruling AD, Task 17b, ead7902)" with the green 5x3 ladder,
what the symptom was, what it was **not** (nothing in `fr-router` was wrong — the
Task 15a/15b off-by-one hypothesis is disproved) and the actual cause; its
obligation row drops the XDIFF; `scripts/differential/README.md`'s `p6t1`
paragraph is corrected and names the new pass-2 stem. The plan's Plan 7 list is
updated too. The three frozen handoffs are deliberately left alone.

S2 — Java's `new Polyline(Line[])` normalises the **caller's** array in place:
`removeConsecutiveParallelLines` (Polyline.java:118) and `removeOverlaps` (:165)
return the input array when they skip nothing, and the constructor then writes
`filteredLines[i] = filteredLines[i].opposite()` (:97) into it. Five Plan 6
tightener sites re-read that array (`TraceTightener.java:297`->`:311`,
`TraceTightener45.java:421`->`:435`, `TraceTightenerAnyAngle.java:158`->`:186`,
`:451`->`:465`, `:614`->`:625`), so since quirk #74 the *identity* of what they
read back is board-observable. Reproduced: `Polyline::from_lines_in_place`
performs the write-back exactly when Java's aliasing holds — decidable without
modelling aliasing, since a normaliser that skips anything returns a strictly
shorter copy — and `new_polyline_in_place` wraps it for the five sites. Quirk
**#188**; quirk #74's "the port reusing where Java rebuilds cannot happen" is
corrected. Two new `fr-geometry` tests, the first of which fails on the old
copy-in behaviour.

S3 + S4 + N3 (controller **ruling AE**) — the identity token is ACCEPTED as the
port of Java object identity, and the departure from the "no static mutable
state" Global Constraint is now declared at the `Line::identity` field, in
`docs/java-quirks.md` §Process notes, in the plan's amendment block and on the
constraint line itself. The counter is widened to **`AtomicU64`**: the port mints
tokens Java never allocates (one per `Polyline::from_lines`, one per polyline
segment inside every `check_trace_shape`), and a wrap would be a silent false
"same object". One reader, one caller, output never depends on the value.

S5 — `// Java bug:` markers at `Board::change_trace`'s two loops (each naming
`PolylineTrace.change` and its Java line on the marker line) and at the
`identity` field. `grep -rn "// Java bug:"` now finds quirk #74.

S6 — `change_keeps_the_entries_whose_lines_are_the_same_objects` is now
discriminating: it rebuilds one line with an unchanged *value* and changes
another for real, so identity gives `keepAtStartCount = 0` and a value comparison
would keep two leaves. With the two comparisons reverted to `!=`, both `change_*`
tests fail where before only one did.

S7 — `scripts/differential/java/p6t17b-bisect.patch` (496 lines) commits levels
2-6 of the bisect. Verified: applies with `patch -p1`; the patched classes with
markers off reproduce `p6t1` byte for byte; and `P6T17B_CHANGE=1` reproduces the
pinning measurement — 2358 `CHANGE` lines, 1403 disagreements — from the
committed tree. `P6T17bProbe` gained `P6T17B_DEEP` / `P6T17B_TREE` /
`P6T17B_CHANGE` and reports clearly when the patch is absent.

S8 — `p2t11` mode 8 gains **S17**: `change` to a value-equal but freshly built
polyline. The jar leaves one trace `[(0,0) (30000,0)]` where a value-comparing
port leaves two, so quirk #74's control-flow half now has a live Java oracle
instead of a prose record.

N1 `Line` is 16 -> 24 bytes (u64, no padding). N2 the "a `Vec` panics the same
way" comments now say **out of bounds only**; an unwritten in-bounds slot carries
the filler's value and token where Java NPEs (unreachable at all three sites, and
each says so). N4 the transcript header carries the `javac` line, the
co-compilation with `P6T1.java`, and the flag-independence note.

Verification (exit column):

| command | result | exit |
|---|---|---|
| cargo fmt --all | clean | 0 |
| cargo clippy --workspace --all-targets -- -D warnings | clean | 0 |
| cargo test --workspace | 1859 passed / 0 failed / 7 ignored | 0 |
| cargo test --workspace --release | 1864 passed / 0 failed / 1 ignored | 0 |
| cargo doc -p fr-geometry / fr-board / fr-router --no-deps | 0 / 1 / 3 warnings (baseline) | 0 |
| audit-port.sh autoroute{,/maze,/expansion,/drill,/path} + fr-router.map | 0 MISSING / 0 UNMAPPED each | 0 |
| audit-port.sh board/actions (3 files), board/optimize TraceShover.java | 0 / 0 | 0 |
| audit-geometry-port.sh | clean | 0 |
| run.sh p2t11 6 / 7 / 8 / 9 / 10 | MATCH 24 / 35 / 49 / 35 / 5 | 0 |
| run.sh p2t15 42 30 | MATCH 521 | 0 |
| run.sh p2t3 / p2t3r / p2t10 / p2t13 | MATCH 132 / 109940 / 88 / 141 | 0 |
| run.sh p3t2 / p3t3 / p3t15 / d17 | MATCH 100000 / 80308 / 378 / 200 | 0 |
| run.sh p4t1 / p5t1 / p5t2 | MATCH 5728 / 937 / 4 | 0 |
| run.sh p6t2 | MATCH 26979 | 0 |
| run.sh p6t3 5 42 20 1000 | MATCH 28812 | 0 |
| run.sh t14 / t15 / t16r / e15 | DIFF — documented; port output byte-identical to b6bb417 | 1 |
| run.sh p6t1 rpi-splitter k=1..8 | MATCH, all eight | 0 |
| run.sh p6t1 dac2020 267 2 / 270 2 | MATCH 268 / 271 | 0 |
| run.sh p6t1 rpi-splitter 8, passes 1/2/4 | MATCH 9 / 9 / 9 | 0 |
| run.sh p6t1 dac2020-bm01 100000, passes 1/2/4 | MATCH 295 / 295 / 295 | 0 |
| run.sh p6t1 j2-reference 100000, passes 1/2/4 | MATCH 46 / 46 / 46 | 0 |
| run.sh p6t1 tutorial-board 8, passes 1/2/4 | MATCH 1 / 1 / 1 | 0 |
| run.sh p6t1 ecc83-input 100000, passes 1/2/4 | MATCH 23 / 23 / 23 | 0 |
| cargo test -p fr-router --test reference_parity --release | 9 passed / 0 failed | 0 |
| gen-router-reference.sh --verify-hash-modes | 6 stems, 5 hash modes agree each | 0 |
| P6T9 / P6T10b / P6T15a / P6T15b / P6T16 probes recompiled | all five transcripts byte-identical | 0 |
| p6t17b-bisect.patch | applies -p1; markers off reproduce p6t1 byte for byte; 2358/1403 reproduced | 0 |

+++ crates/fr-geometry/src/line.rs
@@ -8,7 +8,7 @@
 //! which turns Java's warning into a compile-time guarantee.
 
 use std::cmp::Ordering;
-use std::sync::atomic::{AtomicU32, Ordering as AtomicOrdering};
+use std::sync::atomic::{AtomicU64, Ordering as AtomicOrdering};
 
 use num_bigint::BigInt;
 use num_integer::Integer;
@@ -47,27 +47,66 @@ pub struct Line {
     pub b: IntPoint,
     /// Java's **object identity**, and nothing else — see [`Line::is_same_object`].
     ///
+    // Java bug: PolylineTrace.change (PolylineTrace.java:960, :972) compares two `Line` objects
+    // with `!=` where it means `equals`; this field is the mechanism that reproduces it. Quirk
+    // #74; the reproduction sites are `Board::change_trace`'s two loops in
+    // `crates/fr-board/src/board/trace_normalize.rs`.
+    ///
     /// Java's `Line` is a heap object, and one caller compares two of them with `!=` rather than
     /// `equals`: `PolylineTrace.change` (PolylineTrace.java:960, :972) walks the old and the new
     /// polyline looking for the first and the last line that is *not the same object*, and the
     /// two indices decide how many search-tree leaves the changed trace reuses
     /// (`ShapeSearchTree.changeEntries`). A value comparison answers a different question and
     /// keeps more leaves, which leaves the search tree a different *shape* — see
-    /// `docs/java-quirks.md` and `Board::change_trace`.
+    /// `docs/java-quirks.md` quirk #74 and `Board::change_trace`.
     ///
     /// This port's `Line` is a `Copy` value, so it carries the identity itself: a fresh
     /// constructor call takes a fresh token, exactly as `new Line(...)` allocates a fresh
     /// object, and every copy of the value keeps the token, exactly as copying a Java reference
     /// keeps the object. The token takes no part in `PartialEq`, `Eq`, `Hash` or `Debug`, so no
     /// other comparison in the port can see it.
-    identity: u32,
+    ///
+    /// # Plan 6 controller ruling AE — an accepted departure from "no static mutable state"
+    ///
+    /// The token comes from a process-wide counter, and the plan's Global Constraints forbid
+    /// static mutable state (`docs/superpowers/plans/2026-08-29-plan-6-router-maze.md`, Global
+    /// Constraints, and the plan's amendment block, which records this ruling). Ruling AE
+    /// **accepts** it as the port of Java object identity, for three reasons:
+    ///
+    /// 1. **Nothing reads the counter's value — only the equivalence relation it induces.**
+    ///    `identity` has exactly one reader in the workspace ([`Line::is_same_object`]), which
+    ///    has exactly two callers, both inside one `Board::change_trace` call. No output,
+    ///    ordering, hash or serialised form can observe a token, so two runs over the same input
+    ///    route identically however many `Line`s the process built before.
+    /// 2. **Java's own object identity is process-global mutable state.** A monotone counter is
+    ///    the closest available model, not an invention; the constraint exists to stop the crate
+    ///    inventing a threading policy, and this invents none.
+    /// 3. **Every alternative is worse.** An arena of lines makes `Line::opposite`/`translate`/
+    ///    `turn_90_degree`/`mirror_*` take it as a parameter — a workspace-wide API rewrite for
+    ///    the same mutable state, merely passed explicitly. `Rc`/`Arc<LineData>` pointer identity
+    ///    kills `Copy`, costs an allocation per `Line` in the router's hottest loops, and reuses
+    ///    freed addresses (worse than a counter for silent collisions). A `thread_local!` counter
+    ///    reintroduces cross-thread collisions on a `Send + Sync` `Board`.
+    ///
+    /// The counter is [`u64`] rather than `u32` (ruling AE, review finding S4): the port mints
+    /// tokens Java never allocates — `Polyline::from_polygon`'s overwritten placeholder,
+    /// `remove_overlaps`' filler (one per `Polyline::from_lines`) and
+    /// `offset_shapes_between`'s `[Line; 4]` filler (**one per polyline segment**, inside every
+    /// `check_trace_shape`) — so a `u32` would wrap after enough routing in one process, and a
+    /// wrap is a *silent* false "same object": a wrong `keepAtStartCount`, a different tree
+    /// shape, a different route, no crash. Plan 8 adds a long-lived server process. At
+    /// 2^64 the counter cannot wrap in any run this program can perform.
+    identity: u64,
 }
 
-/// The token source behind `Line::identity`. `Relaxed` is enough: the tokens are only ever
-/// compared for equality, never ordered, printed or persisted.
-static LINE_IDENTITY: AtomicU32 = AtomicU32::new(1);
+/// The token source behind `Line::identity` (Plan 6 controller ruling AE — see that field).
+///
+/// `Relaxed` is enough: the tokens are only ever compared for equality, never ordered, printed or
+/// persisted, so no happens-before relationship is being published. `u64` cannot wrap in any run
+/// this program can perform.
+static LINE_IDENTITY: AtomicU64 = AtomicU64::new(1);
 
-fn next_line_identity() -> u32 {
+fn next_line_identity() -> u64 {
     LINE_IDENTITY.fetch_add(1, AtomicOrdering::Relaxed)
 }
 
@@ -99,7 +138,8 @@ impl std::fmt::Debug for Line {
 impl Line {
     /// Java's `lineA != lineB` — **reference** identity, not the geometric `equals`.
     ///
-    /// True only for a value copied from the same constructor call; see the private `identity` field.
+    /// True exactly for a value copied from the same constructor call; the `u64` token cannot
+    /// wrap, so there is no false positive (see the private `identity` field and ruling AE).
     /// The one caller is `Board::change_trace` (`PolylineTrace.change`, PolylineTrace.java:960
     /// and :972); nothing else in the port may use it to stand in for `==`.
     pub fn is_same_object(&self, other: &Line) -> bool {
